Fix go generate
