+++ tavern/internal/graphql/gqlgen.yml
@@ -24,14 +24,14 @@ resolver:
 # gqlgen will search for any type names in the schema in the generated
 # ent package. If they match it will use them, otherwise it will new ones.
 autobind:
-  - github.com/kcarretto/realm/tavern/ent
-  - github.com/kcarretto/realm/tavern/ent/file
-  - github.com/kcarretto/realm/tavern/ent/user
-  - github.com/kcarretto/realm/tavern/ent/tag
-  - github.com/kcarretto/realm/tavern/ent/beacon
-  - github.com/kcarretto/realm/tavern/ent/quest
-  - github.com/kcarretto/realm/tavern/ent/task
-  - github.com/kcarretto/realm/tavern/ent/tome
+  - github.com/kcarretto/realm/tavern/internal/ent
+  - github.com/kcarretto/realm/tavern/internal/ent/file
+  - github.com/kcarretto/realm/tavern/internal/ent/user
+  - github.com/kcarretto/realm/tavern/internal/ent/tag
+  - github.com/kcarretto/realm/tavern/internal/ent/beacon
+  - github.com/kcarretto/realm/tavern/internal/ent/quest
+  - github.com/kcarretto/realm/tavern/internal/ent/task
+  - github.com/kcarretto/realm/tavern/internal/ent/tome
 
 struct_tag: json
 
@@ -42,13 +42,13 @@ models:
   Node:
     model:
       # ent.Noder is the new interface generated by the Node template.
-      - github.com/kcarretto/realm/tavern/ent.Noder
+      - github.com/kcarretto/realm/tavern/internal/ent.Noder
       # - entgo.io/contrib/entgql/internal/todo/ent.Noder
 
   # Enums
   Kind:
     model:
-      - github.com/kcarretto/realm/tavern/ent/tag.Kind
+      - github.com/kcarretto/realm/tavern/internal/ent/tag.Kind
   HostPlatform:
     model:
-      - github.com/kcarretto/realm/tavern/ent/beacon.HostPlatform
+      - github.com/kcarretto/realm/tavern/internal/ent/beacon.HostPlatform
