--- tavern/internal/www/generate.go
@@ -1,3 +1,0 @@
-package www
-
-//go:generate /bin/sh -c "cat ../../graphql/schema/* > schema.graphql"
